Removed weblab-specific code.
