--- chaste_codegen/_weblab_printer.py
@@ -1,14 +1,0 @@
-from cellmlmanip.printer import Printer
-
-
-class WebLabPrinter(Printer):
-    """
-    Cellmlmanip ``Printer`` subclass for the Web Lab.
-    """
-
-    def __init__(self, symbol_function=None, derivative_function=None):
-        super().__init__(symbol_function, derivative_function)
-
-        # Deal with _exp function introduced to avoid simplification
-        self._function_names['_exp'] = 'math.exp'
-
