--- chaste_codegen/data/tests/conflicting_names.cellml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<model name="generated_model" xmlns="http://www.cellml.org/cellml/1.0#" xmlns:cellml="http://www.cellml.org/cellml/1.0#">
-	<component name="env">
-		<variable name="time" units="dimensionless"/>
-		<math xmlns="http://www.w3.org/1998/Math/MathML">
-			<apply>
-				<eq/>
-				<ci>time</ci>
-				<cn cellml:units="dimensionless">0.0</cn>
-			</apply>
-		</math>
-	</component>
-	<component name="x">
-		<variable name="a" units="dimensionless"/>
-		<variable name="b" units="dimensionless"/>
-		<variable name="y__z" units="dimensionless"/>
-		<math xmlns="http://www.w3.org/1998/Math/MathML">
-			<apply>
-				<eq/>
-				<ci>a</ci>
-				<cn cellml:units="dimensionless">10</cn>
-			</apply>
-			<apply>
-				<eq/>
-				<ci>b</ci>
-				<cn cellml:units="dimensionless">100</cn>
-			</apply>
-			<apply>
-				<eq/>
-				<ci>y__z</ci>
-				<cn cellml:units="dimensionless">1.0</cn>
-			</apply>
-		</math>
-	</component>
-	<component name="x__y">
-		<variable name="a" units="dimensionless"/>
-		<variable name="z" units="dimensionless"/>
-		<variable name="x__y__z" units="dimensionless"/>
-		<math xmlns="http://www.w3.org/1998/Math/MathML">
-			<apply>
-				<eq/>
-				<ci>a</ci>
-				<cn cellml:units="dimensionless">20</cn>
-			</apply>
-			<apply>
-				<eq/>
-				<ci>z</ci>
-				<cn cellml:units="dimensionless">2.0</cn>
-			</apply>
-			<apply>
-				<eq/>
-				<ci>x__y__z</ci>
-				<cn cellml:units="dimensionless">2.0</cn>
-			</apply>
-        </math>
-	</component>
-	<component name="z">
-		<variable name="a" units="dimensionless"/>
-		<variable name="y__z" units="dimensionless"/>
-		<math xmlns="http://www.w3.org/1998/Math/MathML">
-		    <apply>
-			    <eq/>
-			    <ci>a</ci>
-			    <cn cellml:units="dimensionless">3.0</cn>
-		    </apply>
-			<apply>
-				<eq/>
-				<ci>y__z</ci>
-				<cn cellml:units="dimensionless">3.0</cn>
-			</apply>
-        </math>
-	</component>
-</model>
