+++ chaste_codegen/templates/normal_model.cpp
