--- chaste_codegen/templates/wl/TestClampingToInitialValue.pyx
@@ -1,135 +1,0 @@
-# cython: profile=True
-## @file
-## 
-## This source file was generated from CellML.
-## 
-## Model: beeler_reuter_model_1977
-## 
-## Processed by pycml - CellML Tools in Python
-##     (translators: , pycml: , optimize: )
-## on Fri May 18 13:38:04 2018
-## 
-## <autogenerated>
-
-import numpy as np
-
-import fc.simulations.model as Model
-import fc.utility.environment as Env
-import fc.language.values as V
-
-cimport libc.math as math
-cimport numpy as np
-import os
-import shutil
-import sys
-
-from fc.sundials.solver cimport CvodeSolver
-cimport fc.sundials.sundials as Sundials
-from fc.utility.error_handling import ProtocolError
-
-cdef int _EvaluateRhs(Sundials.realtype var_environment__time, Sundials.N_Vector y, Sundials.N_Vector ydot, void* user_data):
-    model = <object>user_data
-    cdef np.ndarray[Sundials.realtype, ndim=1] parameters = <np.ndarray>model.parameters
-    # State variables
-    
-    
-    # Mathematics
-    
-
-cdef class GeneratedModel_Proto_tmpQ4gZsC(CvodeSolver):
-    cdef public char* freeVariableName
-    cdef public double freeVariable
-    cdef public object stateVarMap
-    cdef public np.ndarray initialState
-    cdef public object parameterMap
-    cdef public np.ndarray parameters
-    cdef public object outputNames
-    
-    cdef public object savedStates
-    cdef public object env
-    cdef public bint dirty
-    cdef public char* outputPath
-    cdef public object indentLevel
-    
-    cdef public object _module
-    cdef public object simEnv
-    
-    cdef Sundials.N_Vector _parameters
-    cdef public object _outputs
-    
-    def __init__(self):
-        self.freeVariableName = "time"
-        self.freeVariable = 0.0
-        self.state = np.zeros(0)
-        self.stateVarMap = {}
-        self.initialState = np.zeros(0)
-        
-        self.parameterMap = {}
-        self.parameters = np.zeros(1)
-        self.parameterMap["membrane_voltage"] = 0
-        self.parameters[0] = -84.624 # (c,membrane__V) mV
-        
-        self.outputNames = []
-        outputs = self._outputs = []
-        self.outputNames.append("membrane_voltage")
-        outputs.append(np.array(0.0))
-        self.outputNames.append("time")
-        outputs.append(np.array(0.0))
-        
-        self.state = self.initialState.copy()
-        self.savedStates = {}
-        self.dirty = False
-        self.indentLevel = 0
-        self.AssociateWithModel(self)
-        self._parameters = Sundials.N_VMake_Serial(len(self.parameters), <Sundials.realtype*>(<np.ndarray>self.parameters).data)
-        self.env = Env.ModelWrapperEnvironment(self)
-    
-    def SetRhsWrapper(self):
-        flag = Sundials.CVodeInit(self.cvode_mem, _EvaluateRhs, 0.0, self._state)
-        self.CheckFlag(flag, "CVodeInit")
-    
-    def __dealloc__(self):
-        if self._parameters != NULL:
-            Sundials.N_VDestroy_Serial(self._parameters)
-    
-    def SetOutputFolder(self, path):
-        if os.path.isdir(path) and path.startswith('/tmp'):
-            shutil.rmtree(path)
-        os.mkdir(path)
-        self.outputPath = path
-    
-    def SetIndentLevel(self, indentLevel):
-        self.indentLevel = indentLevel
-    
-    def SetSolver(self, solver):
-        print >>sys.stderr, "  " * self.indentLevel, "SetSolver: Models implemented using Cython contain a built-in ODE solver, so ignoring setting."
-    
-    def GetEnvironmentMap(self):
-        return {'pycml': self.env, 'cmeta': self.env, 'cg': self.env, 'csub': self.env, 'cs': self.env, 'oxmeta': self.env, 'lut': self.env, 'proto': self.env, 'None': self.env, 'bqs': self.env, 'pe': self.env, 'dcterms': self.env, 'xml': self.env, 'dc': self.env, 'bqbiol': self.env, 'cml': self.env, 'solver': self.env, 'doc': self.env, 'm': self.env, 'rdf': self.env, 'cellml': self.env, 'vCard': self.env}
-    
-    cpdef SetFreeVariable(self, double t):
-        self.freeVariable = t
-        CvodeSolver.SetFreeVariable(self, t)
-    
-    def SaveState(self, name):
-        self.savedStates[name] = self.state.copy()
-    
-    cpdef ResetState(self, name=None):
-        if name is None:
-            CvodeSolver.ResetSolver(self, self.initialState)
-        else:
-            CvodeSolver.ResetSolver(self, self.savedStates[name])
-    
-    cpdef GetOutputs(self):
-        cdef np.ndarray[Sundials.realtype, ndim=1] parameters = self.parameters
-        cdef double var_environment__time = self.freeVariable
-        # State variables
-        
-        # Mathematics computing outputs of interest
-        
-        outputs = self._outputs
-        outputs[0][()] = parameters[0]
-        outputs[1][()] = var_environment__time
-        return outputs
-    
-
