--- chaste_codegen/tests/test_weblab_templating.py
@@ -1,70 +1,0 @@
-#
-# Tests templating functionality for the Cardiac Electrophysiology Web Lab
-#
-import cellmlmanip
-import logging
-import os
-import re
-import chaste_codegen as cg
-
-
-# Show more logging output
-logging.getLogger().setLevel(logging.INFO)
-
-
-def test_generate_weblab_model(tmp_path):
-    # Tests the create_weblab_model() method
-
-    # Select output path (in temporary dir)
-    path = tmp_path / 'model.pyx'
-
-    # Select class name
-    class_name = 'TestModel'
-
-    # Load cellml model
-    model = os.path.join(
-        cg.DATA_DIR, 'tests',
-        'hodgkin_huxley_squid_axon_model_1952_modified.cellml'
-    )
-    model = cellmlmanip.load_model(model)
-
-    # Select model outputs (as qualified names)
-    oxmeta = 'https://chaste.comlab.ox.ac.uk/cellml/ns/oxford-metadata#'
-    outputs = [
-        (oxmeta, 'membrane_fast_sodium_current'),
-        (oxmeta, 'membrane_voltage'),
-        (oxmeta, 'time'),
-        'state_variable',
-    ]
-
-    # Select model parameters (as qualified names)
-    parameters = [
-        (oxmeta, 'membrane_fast_sodium_current_conductance'),
-        (oxmeta, 'membrane_potassium_current_conductance'),
-    ]
-
-    # Create weblab model at path
-    cg.create_weblab_model(str(path), class_name, model, outputs, parameters)
-
-    # Read expected output from file
-    expected = os.path.join(cg.DATA_DIR, 'tests', 'weblab_model.pyx')
-    with open(expected, 'r') as f:
-        expected = f.read()
-
-    # Read generated output from file
-    generated = path.read_text()
-
-    # Store locally to update test output file
-    if os.environ.get('WEBLAB_REGENERATE_REF'):
-        print('WRITING TEMPLATING OUTPUT TO FILE')
-        with open('./weblab_model.pyx', 'w') as f:
-            f.write(generated)
-
-    # Remove line about creation date and version
-    p = re.compile('# Generated by .+')
-    expected = p.sub('# Generated by me', expected).strip()
-    generated = p.sub('# Generated by me', generated).strip()
-
-    # Now they should match
-    assert generated == expected
-
